Add a specific error message for when the game is not running
This takes advantage of the updates in memory-rs which return Result types to provide a more user-friendly message if finding the running game fails.

Firstly, it updates main.rs to only compile in the module for the specific game, and then to print the error and exit with an error code.

Secondly, both the zero and kiwami2 modules have been updated to return a Result, and to map the OS error 18 (ERROR_NO_MORE_FILES) to a helpful error message about the game not currently running.

Now, instead of a panic, you get "Error: Yakuza 0 needs to be running"!