If launched outside cmd.exe, wait for input before exiting after errors
This became a bit of a sidetrack for me... 😅

I'm not much of a Windows developer, but I wondered if there was a way to make this not immediately exit when opened from Firefox or Explorer. Eventually I found the `dont_disappear` crate, but it bugged me that it didn't care _where_ the program was launched from.

I ended up wrapping it with a function intended to do its best to determine whether the parent process was `cmd.exe` or `cargo`, and if it's neither it will wait for input. This introduces a few new dependencies, but compiled with msvc the exe still seems pretty small, and I've done my best to ensure this ignores any errors produced while trying to determine this.

Totally understandable if this isn't something you want, but it feels a bit nicer to me, and either way was a nice exercise in figuring out some Windows stuff for me 😇